+++ common/src/main/java/net/arna/jcraft/common/entity/stand/TheSunEntity.java
@@ -550,10 +550,10 @@ public final class TheSunEntity extends StandEntity<TheSunEntity, TheSunEntity.S
         return entityData.get(SCALE);
     }
 
-    @Override
-    public float getScale() {
-        return entityData.get(SCALE);
-    }
+    //@Override
+    //public float getScale() {
+    //    return entityData.get(SCALE);
+    //}
 
     public float getScale(float tickDelta) {
         return Mth.lerp(tickDelta, prevScale, curScale);