+++ doc/changelog.md
@@ -1,8 +1,10 @@
 # Changelog
 ## General
+* __Updated to the newest AzureLib version (3.1.2)__
 * added a hidden advancement for the real hardcore players
 * players can no longer activate end portals in D4C's alt dimension
 ### Blocks & Items
+* Added skins to Anubis Silver Chariot
 * added Jotaro Cosplay from Part 6 (thanks Bob)
 * Diego outfit has been updated
 ### NPCs & Stands
@@ -29,7 +31,11 @@
 * Training Dummy no longer has doubled knockback
 * some cosplay nitpicks
 ### Known Bugs
-* …
+* Some items are duplicated in the Creative Menu
+* Silver Chariot's rapier doesn't render in first-person
+
+
+
 
 ## TODO (SOME UPDATE) :D
 * Spin
@@ -37,5 +43,4 @@
 * MR barrage fire :)
 * Timestop should stop stand anims
 * CRAZY DIAMOND, Hermit Purple, Yellow Temperance
-* Stand NBT serialization
 * Actually use effect keyframes in animations